requirements.txt
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
Cerberus==1.3.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Faker==4.1.0
Flask==1.1.2
Flask-Cors==3.0.8
flask-swagger-ui==3.25.0
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.0.1
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
mysqlclient==1.4.6
packaging==20.4
passlib==1.7.2
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
text-unidecode==1.3
tmdbv3api==1.5.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
